Add support for null-values
An option "nullValue" (string) has been added as an extra argument to
the 2 constructor overloads with the most arguments.
A default value of null is supplied for backwards compatibility. In that
case IsDbNull will still return true when the field "IsNullOrEmpty".
When a nullValue is supplied, IsDbNull will return true when the value
equals (case insensitive) nullValue.
The configured "NullValue" is also exposed as a public readonly
property.